requirements.txt
mujoco
loguru
pyyaml
scipy
torch
onnxruntime
pynput
ipdb
termcolor
sshkeyboard
pygame
zmq
joblib
huggingface-hub
redis
